add reducer, switch case microrefactor
